pyproject.toml
[tool.poetry]
name = "stepic-project-diplom"
version = "0.1.0"
description = ""
authors = ["NIK <nikita1z@mail.ru>"]
readme = "README.md"
packages = [{ include = "stepic_project_diplom" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-environ = "^0.10.0"
of = "^1.0.1"
ruamel-yaml = "^0.17.32"
build = "^0.10.0"
pyproject-flake8 = "^6.0.0.post1"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tools.isort]
py_version = 3.10 # версия питона
profile = 'django'
line_length = 120
multi_line_output = 2 # настройка межстроковых отступов
known_first_party = ["todolist", "core"] # какие приложения являються нами написанными
